fix: move components to features/global
