+++ packages/app/src/features/global/index.ts
@@ -1,2 +1,3 @@
 export { default as SEO } from './SEO'
 export { default as Loading } from './Loading'
+export { default as NotFound } from './NotFound'
