+++ packages/app/src/pages/register.tsx
@@ -1,4 +1,4 @@
-import { SEO } from '@components'
+import { SEO } from '@features/global'
 import useLoggedIn from '@features/auth/hook/useLoggedIn'
 import { RegisterForm } from '@features/register/organisms'
 import RegisterLayout from '@layouts/RegisterLayout'
